Rework get-snapshots predicates
The inner `SnapshotPredicates` class actually only deals with the
`?from_sort_value` filter, and the related `?after` filter was
implemented within `SnapshotSortKey` itself. As a step towards adding a
pre-flight `?after` check this commit refactors the `?after`
functionality into a new `AfterPredicates` class, and for the sake of
consistency moves the `?from_sort_value` functionality to an adjacent
top-level `FromSortValuePredicates` class.
