+++ server/src/main/java/org/elasticsearch/action/admin/cluster/snapshots/get/SnapshotSortKey.java
@@ -9,20 +9,10 @@
 
 package org.elasticsearch.action.admin.cluster.snapshots.get;
 
-import org.elasticsearch.common.io.stream.StreamInput;
-import org.elasticsearch.common.io.stream.StreamOutput;
-import org.elasticsearch.common.io.stream.Writeable;
-import org.elasticsearch.core.Nullable;
-import org.elasticsearch.core.Predicates;
 import org.elasticsearch.search.sort.SortOrder;
 import org.elasticsearch.snapshots.SnapshotInfo;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.Comparator;
-import java.util.function.Predicate;
-import java.util.function.ToLongFunction;
 
 /**
  * Sort key for snapshots e.g. returned from the get-snapshots API. All values break ties using {@link SnapshotInfo#snapshotId} (i.e. by
@@ -32,126 +22,37 @@ public enum SnapshotSortKey {
     /**
      * Sort by snapshot start time.
      */
-    START_TIME("start_time", Comparator.comparingLong(SnapshotInfo::startTime)) {
-        @Override
-        protected String getSortKeyValue(SnapshotInfo snapshotInfo) {
-            return Long.toString(snapshotInfo.startTime());
-        }
-
-        @Override
-        protected Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder) {
-            return after.longValuePredicate(SnapshotInfo::startTime, sortOrder);
-        }
-    },
+    START_TIME("start_time", Comparator.comparingLong(SnapshotInfo::startTime)),
 
     /**
      * Sort by snapshot name.
      */
-    NAME("name", Comparator.comparing(sni -> sni.snapshotId().getName())) {
-        @Override
-        protected String getSortKeyValue(SnapshotInfo snapshotInfo) {
-            return snapshotInfo.snapshotId().getName();
-        }
-
-        @Override
-        protected Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder) {
-            // TODO: cover via pre-flight predicate
-            final String snapshotName = after.snapshotName();
-            final String repoName = after.repoName();
-            return sortOrder == SortOrder.ASC
-                ? (info -> compareName(snapshotName, repoName, info) < 0)
-                : (info -> compareName(snapshotName, repoName, info) > 0);
-        }
-    },
+    NAME("name", Comparator.comparing(sni -> sni.snapshotId().getName())),
 
     /**
      * Sort by snapshot duration (end time minus start time).
      */
-    DURATION("duration", Comparator.comparingLong(sni -> sni.endTime() - sni.startTime())) {
-        @Override
-        protected String getSortKeyValue(SnapshotInfo snapshotInfo) {
-            return Long.toString(snapshotInfo.endTime() - snapshotInfo.startTime());
-        }
-
-        @Override
-        protected Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder) {
-            return after.longValuePredicate(info -> info.endTime() - info.startTime(), sortOrder);
-        }
-    },
+    DURATION("duration", Comparator.comparingLong(sni -> sni.endTime() - sni.startTime())),
 
     /**
      * Sort by number of indices in the snapshot.
      */
-    INDICES("index_count", Comparator.comparingInt(sni -> sni.indices().size())) {
-        @Override
-        protected String getSortKeyValue(SnapshotInfo snapshotInfo) {
-            return Integer.toString(snapshotInfo.indices().size());
-        }
-
-        @Override
-        protected Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder) {
-            // TODO: cover via pre-flight predicate
-            return after.longValuePredicate(info -> info.indices().size(), sortOrder);
-        }
-    },
+    INDICES("index_count", Comparator.comparingInt(sni -> sni.indices().size())),
 
     /**
      * Sort by number of shards in the snapshot.
      */
-    SHARDS("shard_count", Comparator.comparingInt(SnapshotInfo::totalShards)) {
-        @Override
-        protected String getSortKeyValue(SnapshotInfo snapshotInfo) {
-            return Integer.toString(snapshotInfo.totalShards());
-        }
-
-        @Override
-        protected Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder) {
-            return after.longValuePredicate(SnapshotInfo::totalShards, sortOrder);
-        }
-    },
+    SHARDS("shard_count", Comparator.comparingInt(SnapshotInfo::totalShards)),
 
     /**
      * Sort by number of failed shards in the snapshot.
      */
-    FAILED_SHARDS("failed_shard_count", Comparator.comparingInt(SnapshotInfo::failedShards)) {
-        @Override
-        protected String getSortKeyValue(SnapshotInfo snapshotInfo) {
-            return Integer.toString(snapshotInfo.failedShards());
-        }
-
-        @Override
-        protected Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder) {
-            return after.longValuePredicate(SnapshotInfo::failedShards, sortOrder);
-        }
-    },
+    FAILED_SHARDS("failed_shard_count", Comparator.comparingInt(SnapshotInfo::failedShards)),
 
     /**
      * Sort by repository name.
      */
-    REPOSITORY("repository", Comparator.comparing(SnapshotInfo::repository)) {
-        @Override
-        protected String getSortKeyValue(SnapshotInfo snapshotInfo) {
-            return snapshotInfo.repository();
-        }
-
-        @Override
-        protected Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder) {
-            // TODO: cover via pre-flight predicate
-            final String snapshotName = after.snapshotName();
-            final String repoName = after.repoName();
-            return sortOrder == SortOrder.ASC
-                ? (info -> compareRepositoryName(snapshotName, repoName, info) < 0)
-                : (info -> compareRepositoryName(snapshotName, repoName, info) > 0);
-        }
-
-        private static int compareRepositoryName(String name, String repoName, SnapshotInfo info) {
-            final int res = repoName.compareTo(info.repository());
-            if (res != 0) {
-                return res;
-            }
-            return name.compareTo(info.snapshotId().getName());
-        }
-    };
+    REPOSITORY("repository", Comparator.comparing(SnapshotInfo::repository));
 
     private final String name;
     private final Comparator<SnapshotInfo> ascendingSnapshotInfoComparator;
@@ -178,59 +79,6 @@ public enum SnapshotSortKey {
         };
     }
 
-    /**
-     * @return an {@link After} which can be included in a {@link GetSnapshotsRequest} (e.g. to be sent to a remote node) and ultimately
-     * converted into a predicate to filter out {@link SnapshotInfo} items which were returned on earlier pages of results. See also
-     * {@link #encodeAfterQueryParam} and {@link #getAfterPredicate}.
-     */
-    public static After decodeAfterQueryParam(String param) {
-        final String[] parts = new String(Base64.getUrlDecoder().decode(param), StandardCharsets.UTF_8).split(",");
-        if (parts.length != 3) {
-            throw new IllegalArgumentException("invalid ?after parameter [" + param + "]");
-        }
-        return new After(parts[0], parts[1], parts[2]);
-    }
-
-    /**
-     * @return an encoded representation of the value of the sort key for the given {@link SnapshotInfo}, including the values of the
-     * snapshot name and repo name for tiebreaking purposes, which can be returned to the user so they can pass it back to the
-     * {@code ?after} param of a subsequent call to the get-snapshots API in order to retrieve the next page of results.
-     */
-    public final String encodeAfterQueryParam(SnapshotInfo snapshotInfo) {
-        final var rawValue = getSortKeyValue(snapshotInfo) + "," + snapshotInfo.repository() + "," + snapshotInfo.snapshotId().getName();
-        return Base64.getUrlEncoder().encodeToString(rawValue.getBytes(StandardCharsets.UTF_8));
-    }
-
-    /**
-     * @return a string representation of the value of the sort key for the given {@link SnapshotInfo}, which should be the last item in the
-     * response, which is combined with the snapshot and repository names, encoded, and returned to the user so they can pass it back to
-     * the {@code ?after} param of a subsequent call to the get-snapshots API in order to retrieve the next page of results.
-     */
-    protected abstract String getSortKeyValue(SnapshotInfo snapshotInfo);
-
-    /**
-     * @return a predicate to filter out {@link SnapshotInfo} items that match the user's query but which sort earlier than the given
-     * {@link After} value (i.e. they were returned on earlier pages of results). If {@code after} is {@code null} then the returned
-     * predicate matches all snapshots.
-     */
-    public final Predicate<SnapshotInfo> getAfterPredicate(@Nullable After after, SortOrder sortOrder) {
-        return after == null ? Predicates.always() : innerGetAfterPredicate(after, sortOrder);
-    }
-
-    /**
-     * @return a predicate to filter out {@link SnapshotInfo} items that match the user's query but which sort earlier than the given
-     * {@link After} value (i.e. they were returned on earlier pages of results). The {@code after} parameter is not {@code null}.
-     */
-    protected abstract Predicate<SnapshotInfo> innerGetAfterPredicate(After after, SortOrder sortOrder);
-
-    private static int compareName(String name, String repoName, SnapshotInfo info) {
-        final int res = name.compareTo(info.snapshotId().getName());
-        if (res != 0) {
-            return res;
-        }
-        return repoName.compareTo(info.repository());
-    }
-
     public static SnapshotSortKey of(String name) {
         return switch (name) {
             case "start_time" -> START_TIME;
@@ -242,30 +90,5 @@ public enum SnapshotSortKey {
             case "repository" -> REPOSITORY;
             default -> throw new IllegalArgumentException("unknown sort key [" + name + "]");
         };
-    }
-
-    public record After(String value, String repoName, String snapshotName) implements Writeable {
-
-        After(StreamInput in) throws IOException {
-            this(in.readString(), in.readString(), in.readString());
-        }
-
-        @Override
-        public void writeTo(StreamOutput out) throws IOException {
-            out.writeString(value);
-            out.writeString(repoName);
-            out.writeString(snapshotName);
-        }
-
-        Predicate<SnapshotInfo> longValuePredicate(ToLongFunction<SnapshotInfo> extractor, SortOrder sortOrder) {
-            final var after = Long.parseLong(value);
-            return sortOrder == SortOrder.ASC ? info -> {
-                final long val = extractor.applyAsLong(info);
-                return after < val || (after == val && compareName(snapshotName, repoName, info) < 0);
-            } : info -> {
-                final long val = extractor.applyAsLong(info);
-                return after > val || (after == val && compareName(snapshotName, repoName, info) > 0);
-            };
-        }
     }
 }
